move frost wand cast goal to own package
